backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pyodbc>=5.1.0
chardet>=5.2.0
PyYAML>=6.0.2
pandas>=2.2.2
dbfread>=2.0.7
dbf>=0.99.3
python-multipart==0.0.20